refactor : moved socket state to app component
